requirements.txt
streamlit
numpy
scipy
pillow
matplotlib
joblib
scikit-learn
gdown
tensorflow
torch
torchvision
